+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.0] - 2018-11-30
+## [1.4.0] - 2018-12-04
 ### Changed
 - Upgraded backend to EasyAdmin 2
 - Sorted and renamed admin templates
 - Added device preview
+- Improved youtube id parsing in youtube element
+- Added add-element modal instead of dropdown
 
 ## [1.3.10] - 2018-11-22
 ### Added